requirements.txt
PyPDF2
streamlit==1.24.0
google.generativeai
python-dotenv